+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All Notable changes to `League\Url` will be documented in this file
 
 ### Added
 
+- A `Parser` class to parse an URI according to RFC3986 rules, the library no longer depend on PHP `parse_url` function
 - A `Data` class to specifically manipulate `data` schemed URI
 - A `Http` class to specifically manipulate `http`,`https` schemed URI
 - A `Ftp` class to specifically manipulate `ftp` schemed URI
@@ -22,7 +23,6 @@ All Notable changes to `League\Url` will be documented in this file
 - `Host::getIpLiteral` to get the raw IP representation of a Ip Literal hostname
 - `getLiteral` method to `Pass`, `User` and `Fragment` objects to get the component non-encoded string representation
 - `HierarchicalPath::relativize` and `Uri::relativize` to generate relative path and uri respectively
-- The library no longer depend on PHP `parse_url` function
 
 ### Fixed